spacetimedb_lib::bsatn

Trait ToBsatn

pub trait ToBsatn {
    // Required methods
    fn to_bsatn_vec(&self) -> Result<Vec<u8>, BsatnError>;
    fn to_bsatn_extend(&self, buf: &mut Vec<u8>) -> Result<(), BsatnError>;
    fn static_bsatn_size(&self) -> Option<u16>;
}

Types that can be encoded to BSATN.

Implementations of this trait may be more efficient than directly calling to_vec. In particular, for spacetimedb_table::table::RowRef, this method will use a StaticLayout if one is available, avoiding expensive runtime type dispatch.

Required Methods

fn to_bsatn_vec(&self) -> Result<Vec<u8>, BsatnError>

BSATN-encode the row referred to by self into a freshly-allocated Vec<u8>.

fn to_bsatn_extend(&self, buf: &mut Vec<u8>) -> Result<(), BsatnError>

BSATN-encode the row referred to by self into buf, pushing self’s bytes onto the end of buf, similar to Vec::extend.

fn static_bsatn_size(&self) -> Option<u16>

Returns the static size of the type of this object.

When this returns Some(_) there is also a StaticLayout.