spacetimedb_lib::connection_id

Struct ConnectionId

pub struct ConnectionId { /* private fields */ }

A unique identifier for a client connection to a SpacetimeDB database.

This is a special type.

A ConnectionId is a 128-bit unsigned integer. This can be serialized in various ways.

In JSON, an ConnectionId is represented as a BARE DECIMAL number. This requires some care when deserializing; see https://stackoverflow.com/questions/69644298/how-to-make-json-parse-to-treat-all-the-numbers-as-bigint

In BSATN, a ConnectionId is represented as a LITTLE-ENDIAN number 16 bytes long.

In memory, a ConnectionId is stored as a 128-bit number with the endianness of the host system.

Implementations

impl ConnectionId

pub const ZERO: Self

pub const fn from_u128(__connection_id__: u128) -> Self

pub const fn to_u128(&self) -> u128

pub const fn from_le_byte_array(arr: [u8; 16]) -> Self

Create an ConnectionId from a little-endian byte array.

If you are parsing an ConnectionId from a string, you probably want Self::from_be_byte_array instead. But if you need to convert a hexadecimal string to a ConnectionId, just use Self::from_hex.

pub const fn from_be_byte_array(arr: [u8; 16]) -> Self

Create an ConnectionId from a big-endian byte array.

This method is the correct choice if you have converted the bytes of a hexadecimal-formatted ConnectionId to a byte array in the following way:

"0xb0b1b2..."
->
[0xb0, 0xb1, 0xb2, ...]

But if you need to convert a hexadecimal string to a ConnectionId, just use Self::from_hex.

pub const fn as_le_byte_array(&self) -> [u8; 16]

Convert a ConnectionId to a little-endian byte array.

pub const fn as_be_byte_array(&self) -> [u8; 16]

Convert a ConnectionId to a big-endian byte array.

This is a format suitable for printing as a hexadecimal string. But if you need to convert a ConnectionId to a hexadecimal string, just use Self::to_hex.

pub fn from_hex(hex: &str) -> Result<Self, Error>

Parse a hexadecimal string into a ConnectionId.

pub fn to_hex(self) -> HexString<16>

Convert this ConnectionId to a hexadecimal string.

pub fn abbreviate(&self) -> [u8; 8]

Extract the first 8 bytes of this ConnectionId as if it was stored in big-endian format. (That is, the most significant bytes.)

pub fn to_abbreviated_hex(self) -> HexString<8>

Extract the first 16 characters of this ConnectionId’s hexadecimal representation.

pub fn from_be_slice(slice: impl AsRef<[u8]>) -> Self

Create an ConnectionId from a slice, assumed to be in big-endian format.

pub fn to_ipv6(self) -> Ipv6Addr

pub fn to_ipv6_string(self) -> String

pub fn none_if_zero(self) -> Option<Self>