spacetimedb_lib::db

Module default_element_ordering

This module defines the default ordering for fields of a ProductType and variants of a SumType.

In ABI version 8, the default ordering was not applied.

In ABI version 9, the default ordering is applied to all types in a spacetime module, unless they explicitly declare a custom ordering.

Enums

ElementLabel

A label for a field of a ProductType or a variant of a SumType.

Functions

product_type_has_default_ordering

Checks if a product type has the default ordering.

sum_type_has_default_ordering

Checks if a sum type has the default ordering.