spacetimedb_lib::db

Module raw_def

Raw definitions of the database schema.

Modules serialize these types and send them across the ABI boundary to describe to the database what tables they expect. (Wrapped in the type spacetimedb_lib::ModuleDef.)

There will eventually be multiple versions of these types wrapped in a top-level enum. This is because the only backwards-compatible schema changes allowed by BSATN is adding variants to an existing enum. The spacetimedb_schema crate will in the future perform validation and normalization of these Raw types to a canonical form, which will be used everywhere.

Re-exports

pub use v8::*;

Modules

v8

Database definitions v8, the last version before they were wrapped in an enum.

v9

ABI Version 9 of the raw module definitions.