spacetimedb_lib::db::raw_def

Module v10

ABI Version 10 of the raw module definitions.

This is a refactored version of V9 with a section-based structure. V10 moves schedules, lifecycle reducers, and default values out of their V9 locations into dedicated sections for cleaner organization. It allows easier future extensibility to add new kinds of definitions.

Structs

RawColumnDefaultValueV10

Marks a particular table column as having a particular default value.

RawConstraintDefV10

A constraint definition attached to a table.

RawIndexDefV10

The definition of a database index.

RawLifeCycleReducerDefV10

A lifecycle reducer assignment.

RawModuleDefV10

A possibly-invalid raw module definition.

RawModuleDefV10Builder

A builder for a RawModuleDefV10.

RawProcedureDefV10

A procedure definition.

RawReducerDefV10

A reducer definition.

RawScheduleDefV10

A schedule definition.

RawScopedTypeNameV10

A scoped type name, in the form scope0::scope1::...::scopeN::name.

RawSequenceDefV10

A sequence definition for a database table column.

RawTableDefBuilderV10

Builder for a RawTableDefV10.

RawTableDefV10

The definition of a database table.

RawTypeDefV10

A type declaration.

RawViewDefV10

A view definition.

Enums

FunctionVisibility

The visibility of a function (reducer or procedure).

RawModuleDefV10Section

A section of a V10 module definition.

Functions

reducer_default_err_return_type

reducer_default_ok_return_type

sats_name_to_scoped_name_v10

Convert a string from a sats type-name annotation like #[sats(name = "namespace.name")] to a RawScopedTypeNameV9. We split the input on the strings "::" and "." to split up module paths.

Type Aliases

RawRowLevelSecurityDefV10