spacetimedb_lib::db::raw_def

Module v9

ABI Version 9 of the raw module definitions.

This is the ABI that will be used for 1.0. We are keeping around the old ABI (v8) for now, to allow ourselves to convert the codebase a component-at-a-time.

Structs

RawColumnDefaultValueV9

Marks a particular table’s column as having a particular default.

RawConstraintDefV9

A constraint definition attached to a table.

RawIndexDefV9

The definition of a database index.

RawModuleDefV9

A possibly-invalid raw module definition.

RawModuleDefV9Builder

A builder for a RawModuleDefV9.

RawProcedureDefV9

A procedure definition.

RawReducerDefV9

A reducer definition.

RawRowLevelSecurityDefV9

Data for the RLS policy on a table.

RawScheduleDefV9

Marks a table as a timer table for a scheduled reducer or procedure.

RawScopedTypeNameV9

A scoped type name, in the form scope0::scope1::...::scopeN::name.

RawSequenceDefV9

A sequence definition for a database table column.

RawTableDefBuilder

Builder for a RawTableDef.

RawTableDefV9

The definition of a database table.

RawTypeDefV9

A type declaration.

RawUniqueConstraintDataV9

Requires that the projection of the table onto these columns is a bijection.

RawViewDefV9

A view definition.

Enums

Lifecycle

Special roles a reducer can play in the module lifecycle.

RawConstraintDataV9

Raw data attached to a constraint. New fields MUST be added to the END of this enum, to maintain ABI compatibility.

RawIndexAlgorithm

Data specifying an index algorithm. New fields MUST be added to the END of this enum, to maintain ABI compatibility.

RawMiscModuleExportV9

A miscellaneous module export.

TableAccess

The visibility of the table.

TableType

Whether the table was created by the system or the user.

ViewResultHeader

Functions

btree

Returns a btree index algorithm for the columns cols.

direct

Returns a direct index algorithm for the column col.

hash

Returns a hash index algorithm for the columns cols.

sats_name_to_scoped_name

Convert a string from a sats type-name annotation like #[sats(name = "namespace.name")] to a RawScopedTypeNameV9. We split the input on the strings "::" and "." to split up module paths.

Type Aliases

RawIdentifier

A not-yet-validated identifier.

RawSql

A not-yet-validated sql.