spacetimedb_lib::db::raw_def::v9

Struct RawTableDefBuilder

pub struct RawTableDefBuilder<'a> { /* private fields */ }

Builder for a RawTableDef.

Implementations

impl RawTableDefBuilder<'_>

pub fn with_type(self, table_type: TableType) -> Self

Sets the type of the table and return it.

This is not about column algebraic types, but about whether the table was created by the system or the user.

pub fn with_access(self, table_access: TableAccess) -> Self

Sets the access rights for the table and return it.

pub fn with_unique_constraint(self, columns: impl Into<ColList>) -> Self

Generates a RawConstraintDefV9 using the supplied columns.

pub fn with_primary_key(self, column: impl Into<ColId>) -> Self

Adds a primary key to the table. You must also add a unique constraint on the primary key column.

pub fn with_auto_inc_primary_key(self, column: impl Into<ColId>) -> Self

Adds a primary key to the table, with corresponding unique constraint and sequence definitions. You will also need to call Self::with_index to create an index on column.

pub fn with_index( self, algorithm: RawIndexAlgorithm, accessor_name: impl Into<RawIdentifier>, ) -> Self

Generates a RawIndexDefV9 using the supplied columns.

pub fn with_index_no_accessor_name(self, algorithm: RawIndexAlgorithm) -> Self

Generates a RawIndexDefV9 using the supplied columns but with no accessor_name.

pub fn with_column_sequence(self, column: impl Into<ColId>) -> Self

Adds a RawSequenceDefV9 on the supplied column.

pub fn with_schedule( self, reducer_name: impl Into<RawIdentifier>, scheduled_at_column: impl Into<ColId>, ) -> Self

Adds a schedule definition to the table.

The table must have the appropriate columns for a scheduled table.

pub fn finish(self) -> AlgebraicTypeRef

Build the table and add it to the module, returning the product_type_ref of the table.

pub fn find_col_pos_by_name(&self, column: impl AsRef<str>) -> Option<ColId>

Get the column ID of the column with the specified name, if any.

Returns None if this TableDef has been constructed with an invalid ProductTypeRef, or if no column exists with that name.