spacetimedb_lib::de

Trait ArrayAccess

pub trait ArrayAccess<'de> {
    type Element;
    type Error: Error;

    // Required method
    fn next_element(&mut self) -> Result<Option<Self::Element>, Self::Error>;

    // Provided method
    fn size_hint(&self) -> Option<usize> { ... }
}

Provides an ArrayVisitor with access to each element of the array in the input.

This is a trait that a Deserializer passes to an ArrayVisitor implementation.

Required Associated Types

type Element

The element / base type of the array.

type Error: Error

The error type that can be returned if some error occurs during deserialization.

Required Methods

fn next_element(&mut self) -> Result<Option<Self::Element>, Self::Error>

This returns Ok(Some(value)) for the next element in the array, or Ok(None) if there are no more remaining elements.

Provided Methods

fn size_hint(&self) -> Option<usize>

Returns the number of elements remaining in the array, if known.