spacetimedb_lib::de

Trait FieldNameVisitor

pub trait FieldNameVisitor<'de> {
    type Output;

    // Required methods
    fn field_names(&self) -> impl Iterator<Item = Option<&str>>;
    fn visit<E>(self, name: &str) -> Result<Self::Output, E>
       where E: Error;
    fn visit_seq(self, index: usize) -> Self::Output;

    // Provided method
    fn kind(&self) -> ProductKind { ... }
}

Visitor used to deserialize the name of a field.

Required Associated Types

type Output

The resulting field name.

Required Methods

fn field_names(&self) -> impl Iterator<Item = Option<&str>>

Provides a list of valid field names.

Where None is yielded, this indicates a nameless field.

fn visit<E>(self, name: &str) -> Result<Self::Output, E>
where E: Error,

Deserializes the name of a field using name.

fn visit_seq(self, index: usize) -> Self::Output

Deserializes the name of a field using index.

Should only be called when index is already known to exist and is expected to panic otherwise.

Provided Methods

fn kind(&self) -> ProductKind

The sort of product deserialized.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.