spacetimedb_lib::db::raw_def::v9

Struct RawModuleDefV9Builder

pub struct RawModuleDefV9Builder { /* private fields */ }

A builder for a RawModuleDefV9.

Implementations

impl RawModuleDefV9Builder

pub fn new() -> Self

Create a new, empty RawModuleDefBuilder.

pub fn add_type<T: SpacetimeType>(&mut self) -> AlgebraicType

Add a type to the in-progress module.

The returned type must satisfy AlgebraicType::is_valid_for_client_type_definition or AlgebraicType::is_valid_for_client_type_use .

pub fn build_table( &mut self, name: impl Into<RawIdentifier>, product_type_ref: AlgebraicTypeRef, ) -> RawTableDefBuilder<'_>

Create a table builder.

Does not validate that the product_type_ref is valid; this is left to the module validation code.

pub fn build_table_with_new_type( &mut self, table_name: impl Into<RawIdentifier>, product_type: impl Into<ProductType>, custom_ordering: bool, ) -> RawTableDefBuilder<'_>

Build a new table with a product type. Adds the type to the module.

pub fn build_table_with_new_type_for_tests( &mut self, table_name: impl Into<RawIdentifier>, product_type: ProductType, custom_ordering: bool, ) -> RawTableDefBuilder<'_>

Build a new table with a product type, for testing. Adds the type to the module.

pub fn add_algebraic_type( &mut self, scope: impl IntoIterator<Item = RawIdentifier>, name: impl Into<RawIdentifier>, ty: AlgebraicType, custom_ordering: bool, ) -> AlgebraicTypeRef

Add a type to the typespace, along with a type alias declaring its name. This method should only be use for AlgebraicTypes not corresponding to a Rust type that implements SpacetimeType.

Returns a reference to the newly-added type.

NOT idempotent, calling this twice with the same name will cause errors during validation.

You must set custom_ordering if you’re not using the default element ordering.

pub fn add_reducer( &mut self, name: impl Into<RawIdentifier>, params: ProductType, lifecycle: Option<Lifecycle>, )

Add a reducer to the in-progress module. Accepts a ProductType of reducer arguments for convenience. The ProductType need not be registered in the typespace.

Importantly, if the reducer’s first argument is a ReducerContext, that information should not be provided to this method. That is an implementation detail handled by the module bindings and can be ignored. As far as the module definition is concerned, the reducer’s arguments start with the first non-ReducerContext argument.

(It is impossible, with the current implementation of ReducerContext, to have more than one ReducerContext argument, at least in Rust. This is because SpacetimeType is not implemented for ReducerContext, so it can never act like an ordinary argument.)

pub fn add_row_level_security(&mut self, sql: &str)

Add a row-level security policy to the module.

The sql expression should be a valid SQL expression that will be used to filter rows.

NOTE: The sql expression must be unique within the module.

pub fn typespace(&self) -> &Typespace

Get the typespace of the module.

pub fn finish(self) -> RawModuleDefV9

Finish building, consuming the builder and returning the module. The module should be validated before use.