spacetimedb_lib::de

Trait SumAccess

pub trait SumAccess<'de> {
    type Error: Error;
    type Variant: VariantAccess<'de, Error = Self::Error>;

    // Required method
    fn variant<V>(
        self,
        visitor: V,
    ) -> Result<(<V as VariantVisitor<'de>>::Output, Self::Variant), Self::Error>
       where V: VariantVisitor<'de>;
}

Provides a SumVisitor access to the data of a sum in the input.

An A: SumAccess object is created by the [D: Deserializer] which passes A to a [V: SumVisitor] that D in turn was passed. A is then used by V to split tag and value input apart.

Required Associated Types

type Error: Error

The error type that can be returned if some error occurs during deserialization.

type Variant: VariantAccess<'de, Error = Self::Error>

The visitor used to deserialize the content of the sum variant.

Required Methods

fn variant<V>( self, visitor: V, ) -> Result<(<V as VariantVisitor<'de>>::Output, Self::Variant), Self::Error>
where V: VariantVisitor<'de>,

Called to identify which variant to deserialize. Returns a tuple with the result of identification (V::Output) and the input to variant data deserialization.

The visitor is provided by the Deserializer. This method is typically called from SumVisitor::visit_sum which will provide the V: VariantVisitor.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.